Implemented huge part of imported items (+ description)
Added and fixed some translations.
Some sounds has been renamed.
The aiming positions of the guns are not set yet for guns other than ACR.
The attachments equipped positions are not set yet for any gun.
Attachments are being rendered as 3D items in the inventory, but they sould render as 2D.
There are new TODO notes in some .java files.